upload keep files for each patient
After receiving cell ranger output we consulted with 10x on which cells to keep and which cells to discard. Granulocytes, or in our case eosinophils, are known to have lower RNA content and express fewer genes compared to other cell types. This will cause these cells to be identified as empty droplets and removed when running Cell Ranger unless we use the ‘force-cells’ argument. Since we aimed for 10,000 cells per lane, we set ‘force-cells=10000’ for each of our Cell Ranger runs. In turn, this kept everything, including empty droplets. It was recommended that remove clusters with low UMI counts. The cells we removed also corresponded to cells expressing very low, or in most cases no expression, of eosinophil markers, indicating that they could be empty droplets or low quality cells.